backend/requirements.txt
# === CORE API ===
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
email-validator==2.3.0
python-multipart==0.0.9

# === DATABASE ===
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# === AUTHENTICATION ===
python-jose==3.3.0
#passlib==1.7.4
#bcrypt==5.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==46.0.3

# === MACHINE LEARNING ===
tensorflow==2.16.1
# NOTE: Do NOT include tensorflow-intel (Windows-only)
numpy==1.26.4
pillow==10.4.0
keras==3.13.1

# === SECURITY & RATE LIMITING ===
slowapi==0.1.9
python-magic==0.4.27

# === OPTIONAL (remove if not needed) ===
matplotlib==3.9.0
rich==14.3.1
requests==2.32.5